@@ -1,28 +1,36 @@
-import { Toolbar, Typography } from '@material-ui/core';
-import { IconButton } from '@material-ui/core';
-import { AppBar } from '@material-ui/core';
 import * as React from 'react';
-import FacebookIcon from '@material-ui/icons/Facebook';
-import GitHubIcon from '@material-ui/icons/GitHub';
-import LinkedInIcon from '@material-ui/icons/LinkedIn';
+import FacebookIcon from '@mui/icons-material/Facebook';
+import GitHubIcon from '@mui/icons-material/GitHub';
+import LinkedInIcon from '@mui/icons-material/LinkedIn';
+import { AppBar, Toolbar, IconButton, Typography } from '@mui/material';
 
-class Footer extends React.Component<{ data: any }, any> {
+interface FooterProps {
+    latest_backup?: string
+}
+
+class Footer extends React.Component<FooterProps, {}> {
     public render() {
         return (
-            <AppBar position="fixed" color="primary" style={{ bottom: 0, top: 'auto' }}>
+            <AppBar position="fixed" style={{ bottom: 0, top: 'auto' }}>
                 <Toolbar variant="dense">
-                    <IconButton edge="start" href="https://www.facebook.com/sanket.revankar1" target="_blank" className='mr-2' size="medium">
-                        <FacebookIcon fontSize='small' color='action' className='text-light' />
+                    <IconButton href="https://www.facebook.com/sanket.revankar1"
+                        target="_blank" className='mr-1'
+                    >
+                        <FacebookIcon fontSize='small' className='text-light' />
                     </IconButton>
-                    <IconButton edge="start" href="https://github.com/SanketRevankar" target="_blank" className='mr-2' size="medium">
-                        <GitHubIcon fontSize='small' color='action' className='text-light' />
+                    <IconButton href="https://github.com/SanketRevankar"
+                        target="_blank" className='mr-1'
+                    >
+                        <GitHubIcon fontSize='small' className='text-light' />
                     </IconButton>
-                    <IconButton edge="start" href="https://www.linkedin.com/in/sanketrevankar/" target="_blank" className='mr-2' size="medium">
-                        <LinkedInIcon fontSize='small' color='action' className='text-light' />
+                    <IconButton href="https://www.linkedin.com/in/sanketrevankar/"
+                        target="_blank" className='mr-1'
+                    >
+                        <LinkedInIcon fontSize='small' className='text-light' />
                     </IconButton>
                     <Typography className='ml-auto'>
-                        {this.props.data.latest_backup ?
-                            'Last Backup on: ' + this.props.data.latest_backup
+                        {this.props.latest_backup ?
+                            'Last Backup on: ' + new Date(parseFloat(this.props.latest_backup) * 1000).toLocaleString('en-IN')
                             :
                             'No backup available.'
                         }
