@@ -1,14 +1,8 @@
-import * as React from 'react';
-import * as ReactDOM from 'react-dom';
+import './index.css';
 import App from './App';
-import * as serviceWorker from './serviceWorker';
+import './bootstrap.min.css'
+import { createRoot } from 'react-dom/client';
 
-ReactDOM.render(
-    <App />,
-    document.getElementById('react') as HTMLElement
-);
-
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.register();
+const container = document.getElementById('root')
+const root = createRoot(container!);
+root.render(<App />);
