@@ -1,4 +1,4 @@
-# Generated by Django 2.2.5 on 2019-09-21 12:10
+# Generated by Django 4.1.6 on 2023-02-12 12:00
 
 from django.db import migrations, models
 
@@ -12,10 +12,13 @@ class Migration(migrations.Migration):
 
     operations = [
         migrations.CreateModel(
-            name='LastBackup',
+            name='Backup',
             fields=[
-                ('id', models.AutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')),
-                ('pub_date', models.TextField(verbose_name='Date last updated')),
+                ('name', models.TextField(primary_key=True, serialize=False, verbose_name='Backup name')),
+                ('location', models.TextField(verbose_name='Backup Location')),
+                ('last_backup_start_time', models.TextField(verbose_name='Date last update started')),
+                ('last_backup_end_time', models.TextField(verbose_name='Date last update ended')),
+                ('config', models.TextField(verbose_name='Backup Config')),
             ],
         ),
     ]
