@@ -1,68 +1,73 @@
+import json
 import logging
-from configparser import ConfigParser, ExtendedInterpolation
 
-from AndroidFTPBackup.constants import PyStrings as pS
-from AndroidFTP_Backup import handler
+from asgiref.sync import sync_to_async
 
-_conf = r'AndroidFTPBackup/conf/'
+from AndroidFTPBackup.utils.FileHelper import FileHelper
 
 
 class ConfigHelper:
+    logger = logging.getLogger(__name__)
+    configs = {}
+    backups = None
 
-    def __init__(self):
-        self.logger = logging.getLogger(__name__)
-        self.logger.info('{} - Initiated'.format(__name__))
-        self.config = self.init_config_parser()
-        self.config.read(_conf + 'main.conf')
-        self.default_loaded = True
-        self.logger.info('Loaded default config')
-        self.backups = eval(self.config[pS.ID_CAPS][pS.BACKUPS])
-        if len(self.backups) != 0:
-            backup_name = self.config[pS.ID_CAPS][pS.DEFAULT_BACKUP]
-            self.load_config(backup_name)
-            self.default_loaded = False
+    @classmethod
+    def load_config(cls, backup_name, init=True, override=False):
+        from AndroidFTPBackup.models import Backup
 
-    def load_config(self, backup_name, init=True):
-        self.logger.info('Loaded config: ' + backup_name)
-        config = self.init_config_parser()
-        config.read(_conf + pS.BACKUPS_ + backup_name)
-        self.config = config
+        if backup_name in cls.configs:
+            return cls.configs[backup_name]
+
+        cls.logger.info('Loading config: {}'.format(backup_name))
+        backup = Backup.objects.filter(name=backup_name).values()[0]
+        cls.configs[backup_name] = backup
+        config = json.loads(backup['config'])
+        backup['config'] = config
+
         if init:
-            handler.fileHelper.async_init()
+            FileHelper.async_init(backup['location'], override)
 
-    def save_config(self, config, backup_name):
-        config_parser = self.init_config_parser()
-        config_parser.read_dict(config)
-        conf_backup_name = _conf + pS.BACKUPS_ + backup_name
-        self.logger.info('Saving config to file: {}'.format(conf_backup_name))
-        self.write_config(conf_backup_name, config_parser)
+        return backup
 
-        main_config = self.init_config_parser()
-        main_config.read(_conf + 'main.conf')
-        backups = eval(main_config[pS.ID_CAPS][pS.BACKUPS])
-        if backup_name not in backups:
-            backups.append(backup_name)
-        main_config[pS.ID_CAPS][pS.BACKUPS] = backups.__str__()
+    @classmethod
+    def load_backups(cls):
+        from AndroidFTPBackup.models import Backup
 
-        if self.default_loaded:
-            main_config[pS.ID_CAPS][pS.DEFAULT_BACKUP] = backup_name
-            self.default_loaded = False
-            self.backups = [backup_name]
+        if cls.backups:
+            return cls.backups
 
-        self.write_config(_conf + 'main.conf', main_config)
-        self.load_config(backup_name)
-        if backup_name not in self.backups:
-            self.backups.append(backup_name)
+        cls.logger.info('Loading all backups')
+        backups = Backup.objects.all().values('name')
+        backups = [backup['name'] for backup in backups]
+        cls.backups = backups
 
-    @staticmethod
-    def write_config(conf_backup_name, config_parser):
-        with open(conf_backup_name, 'w') as configfile:
-            config_parser.write(configfile)
+        return backups
 
-    @staticmethod
-    def init_config_parser():
-        return ConfigParser(allow_no_value=True, inline_comment_prefixes=('#',),
-                            interpolation=ExtendedInterpolation())
+    @classmethod
+    def save_config(cls, config_data):
+        from AndroidFTPBackup.models import Backup
 
-    def get_config(self):
-        return self.config
+        backup_name = config_data['basic']['name']
+        backup_location = config_data['basic']['location']
+        cls.logger.info('Saving config: {}'.format(backup_name))
+        updates = dict(location=backup_location, config=json.dumps(config_data))
+        Backup.objects.update_or_create(name=backup_name, defaults=updates)
+        FileHelper.create_folder(backup_location)
+        cls.clear_backup_config_cache(backup_name)
+
+    @classmethod
+    @sync_to_async
+    def update_backup_time(cls, backup_name, current_backup_start_time, current_backup_end_time):
+        from AndroidFTPBackup.models import Backup
+
+        cls.logger.info('Updating config: {}'.format(backup_name))
+        updates = dict(last_backup_start_time=current_backup_start_time, last_backup_end_time=current_backup_end_time)
+        Backup.objects.update_or_create(name=backup_name, defaults=updates)
+        cls.clear_backup_config_cache(backup_name)
+        cls.load_config(backup_name, True, True)
+
+    @classmethod
+    def clear_backup_config_cache(cls, backup_name):
+        cls.configs.pop(backup_name, None)
+        if cls.backups and backup_name not in cls.backups:
+            cls.backups = None
