@@ -5,76 +5,66 @@ import xml.etree.ElementTree as Et
 
 import nmap
 
-from AndroidFTPBackup.constants import PyStrings as pS
-
 
 class WiFiHelper:
-    def __init__(self):
-        self.logger = logging.getLogger(__name__)
+    logger = logging.getLogger(__name__)
 
-    def get_wifi_connections(self, request):
-        hosts = {}
-        ip = request.GET.get(pS.IP)
-
-        self.logger.info(pS.GET_WIFI_CONNECTIONS.format(ip))
+    @classmethod
+    def get_connected_devices(cls, ip):
+        cls.logger.info('Getting wifi connections on: {}'.format(ip))
         uuid_ = uuid.uuid4().__str__()
-        root = self.start_nmap(ip, uuid_)
+        root = cls.start_nmap(ip, uuid_)
 
-        n = 0
+        devices = []
         for child in root:
-            if child.tag == pS.HOST:
-                for c in child:
-                    if c.tag == pS.ADDRESS:
-                        if c.attrib[pS.ADDRTYPE] == pS.IPV4:
-                            n += 1
-                            hosts[n] = {pS.IP: c.attrib[pS.ADDR]}
-                            self.logger.info(pS.LOG_SEPERATOR)
-                            self.logger.info(pS.FOUND_IP_.format(c.attrib[pS.ADDR]))
-                        if c.attrib[pS.ADDRTYPE] == pS.MAC:
-                            hosts[n][pS.MAC] = c.attrib[pS.ADDR]
-                            self.logger.info(pS.MAC_ID_.format(c.attrib[pS.ADDR]))
-                            if pS.VENDOR in c.attrib:
-                                hosts[n][pS.VENDOR] = c.attrib[pS.VENDOR]
-                                self.logger.info(pS.VENDOR_.format(c.attrib[pS.VENDOR]))
+            if child.tag == 'host':
+                host = {}
+                for element in child:
+                    if element.tag == 'address':
+                        if element.attrib['addrtype'] == 'ipv4':
+                            host['ip'] = element.attrib['addr']
+                        if element.attrib['addrtype'] == 'mac':
+                            host['mac'] = element.attrib['addr']
+                            if 'vendor' in element.attrib:
+                                host['vendor'] = element.attrib['vendor']
+                if 'mac' in host:
+                    devices.append(host)
 
-        self.remove_nmap_temp_file(uuid_)
+        os.remove(uuid_)
 
-        return hosts
+        return devices
 
-    def get_ip_by_mac(self, hosts, mac):
-        self.logger.info(pS.GET_IP_FROM_MAC.format(mac, hosts))
+    @classmethod
+    def get_ip_by_mac(cls, hosts, mac):
+        cls.logger.info('Getting wifi connection with MAC: {} on {}'.format(mac, hosts))
         uuid_ = uuid.uuid4().__str__()
-        root = self.start_nmap(hosts, uuid_)
+        root = cls.start_nmap(hosts, uuid_)
 
         last_ip = ''
         for child in root:
-            if child.tag == pS.HOST:
+            if child.tag == 'host':
                 for c in child:
-                    if c.tag == pS.ADDRESS:
-                        if c.attrib[pS.ADDRTYPE] == pS.IPV4:
-                            last_ip = c.attrib[pS.ADDR]
-                        if c.attrib[pS.ADDRTYPE] == pS.MAC:
-                            if mac == c.attrib[pS.ADDR]:
-                                self.logger.info(pS.WIFI_WITH_MAC_HAS_IP_.format(mac, last_ip))
-                                self.remove_nmap_temp_file(uuid_)
+                    if c.tag == 'address':
+                        if c.attrib['addrtype'] == 'ipv4':
+                            last_ip = c.attrib['addr']
+                        if c.attrib['addrtype'] == 'mac':
+                            if mac == c.attrib['addr']:
+                                cls.logger.info('Wifi connection with MAC: {} has IP: {}'.format(mac, last_ip))
+                                os.remove(uuid_)
                                 return last_ip
 
-        self.remove_nmap_temp_file(uuid_)
+        os.remove(uuid_)
 
-        self.logger.warning(pS.IP_WITH_MAC_NOT_FOUND.format(mac, hosts))
-        return pS.NOT_FOUND
+        cls.logger.warning('Wifi connection with MAC: {} on {} not found'.format(mac, hosts))
+        return 'Not Found'
 
-    @staticmethod
-    def start_nmap(hosts, uuid_):
+    @classmethod
+    def start_nmap(cls, hosts, uuid_):
         nm = nmap.PortScanner()
-        nm.scan(hosts=hosts, arguments=pS.NMAP_ARGS)
+        nm.scan(hosts=hosts, arguments='-sP --max-parallelism 200')
         nm_mxl = nm.get_nmap_last_output()
-        file = open(pS.TEMP_XML + uuid_, pS.WRITE)
+        file = open(uuid_, 'w')
         file.write(nm_mxl)
         file.close()
-        tree = Et.parse(pS.TEMP_XML + uuid_)
+        tree = Et.parse(uuid_)
         return tree.getroot()
-
-    @staticmethod
-    def remove_nmap_temp_file(uuid_):
-        os.remove(pS.TEMP_XML + uuid_)
