@@ -1,12 +1,5 @@
 from django.apps import AppConfig
 
-from AndroidFTPBackup.utils import ConfigHelper, FileHelper
-from AndroidFTP_Backup import handler
 
-
-class MyAppConfig(AppConfig):
+class AndroidFTP_BackupConfig(AppConfig):
     name = 'AndroidFTP_Backup'
-
-    def ready(self):
-        handler.fileHelper = FileHelper.FileHelper()
-        handler.configHelper = ConfigHelper.ConfigHelper()
