@@ -1,10 +1,17 @@
+import os
+
 from channels.auth import AuthMiddlewareStack
 from channels.routing import ProtocolTypeRouter, URLRouter
+from django.core.asgi import get_asgi_application
 
 import AndroidFTPBackup.routing
 
+
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "mysite.settings")
+django_asgi_app = get_asgi_application()
+
 application = ProtocolTypeRouter({
-    # (http->django views is added by default)
+    'http': django_asgi_app,
     'websocket': AuthMiddlewareStack(
         URLRouter(
             AndroidFTPBackup.routing.websocket_urlpatterns
