@@ -1,40 +1,64 @@
-import React, { useState } from 'react';
-import useMediaQuery from '@material-ui/core/useMediaQuery';
-import { createTheme, ThemeProvider } from '@material-ui/core/styles';
-import CssBaseline from '@material-ui/core/CssBaseline';
-import Routes from './Routes';
-import { PaletteType } from '@material-ui/core';
-import { lightBlue, deepPurple, indigo } from '@material-ui/core/colors';
+import './App.css';
+import { useState } from 'react';
+import Main from './Main';
 import { BrowserRouter as Router } from "react-router-dom";
+import { useMediaQuery, createTheme, CssBaseline, ThemeProvider } from '@mui/material';
+import { lightBlue } from '@mui/material/colors';
 
-function App() {
-    const prefersDarkMode = useMediaQuery('(prefers-color-scheme: dark)', { noSsr: true });
-    const savedTheme = localStorage.getItem("dark-mode")
-    const mode: boolean = savedTheme ? savedTheme == 'true' : prefersDarkMode
+export default function App() {
+    const preferedMode = useMediaQuery('(prefers-color-scheme: dark)', { noSsr: true }) ? 'dark' : 'light'
+    const savedMode = localStorage.getItem("mode")
+    const mode = savedMode ? savedMode : preferedMode
+    const [themeType, setTheme] = useState(mode)
+    const darkMode = themeType === 'dark'
 
-    let sTheme: PaletteType = mode ? 'dark' : 'light'
-    const [themeType, setTheme] = useState(sTheme);
     const switchTheme = () => {
-        localStorage.setItem("dark-mode", JSON.stringify(themeType === "light"));
-        setTheme(themeType === "light" ? "dark" : "light");
-    };
+        const newMode = darkMode ? "light" : "dark"
+        setTheme(newMode);
+        localStorage.setItem("mode", newMode);
+    }
 
     const theme = createTheme({
-        palette: {
-            type: themeType,
-            primary: indigo,
-            secondary: themeType === "light" ? deepPurple : lightBlue,
+        palette: darkMode ? {
+            mode: 'dark',
+            primary: {
+                main: 'rgb(144, 183, 249)'
+            },
+            secondary: lightBlue,
+            background: {
+                default: '#252525',
+                paper: '#282828'
+            }
+        } : {
+            mode: 'light',
+            primary: {
+                main: 'rgb(48, 102, 201)'
+            },
+            success: {
+                main: '#00ae07'
+            },
+            background: {
+                default: '#fff',
+                paper: '#f8f8f8'
+            }
         },
+        components: {
+            MuiAppBar: {
+                styleOverrides: {
+                    root: {
+                        backgroundColor: darkMode ? '#011432' : '#1b3052'
+                    }
+                }
+            }
+        }
     })
 
     return (
         <ThemeProvider theme={theme}>
             <CssBaseline />
             <Router basename='/AndroidFTPBackup'>
-                <Routes data={{ switchTheme: switchTheme, themeType: themeType }} />
+                <Main switchTheme={switchTheme} themeType={themeType} />
             </Router>
         </ThemeProvider>
-    );
+    )
 }
-
-export default App;
